fix: rename page-header/main.vue => PageHeader.vue
