+++ packages/page-header/__tests__/pageHeader.spec.js
@@ -1,4 +1,4 @@
-import PageHeader from '../src/main.vue'
+import PageHeader from '../PageHeader.vue'
 import { mount } from '@vue/test-utils'
 
 describe('PageHeader', () => {
